fix: publish step was failing after switch to hybrid packages
Publish step was failing because the prepack script was executed by
lerna in the package's folder.
This is failing for an unknown reason.
Running `yarn workspace package_name build` is working though.

To bypassa the issue the prepack lifecycles has been removed and the
publish job has been moved to the Test_Lint job that has been renamed.

This way we can ensure that the packages are build before deploy.

This way we will always build all packages, whether before we were
building only changed packages. This process is cheap and fast so we
don't bother the trade off just yet.